requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   openai
azure-cognitiveservices-speech==1.42.0
    # via -r requirements.in
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
jiter==0.8.2
    # via openai
linkify-it-py==2.0.3
    # via markdown-it-py
markdown-it-py[linkify,plugins]==3.0.0
    # via
    #   mdit-py-plugins
    #   rich
    #   textual
mdit-py-plugins==0.4.2
    # via markdown-it-py
mdurl==0.1.2
    # via markdown-it-py
numpy==2.2.2
    # via -r requirements.in
openai==1.60.2
    # via -r requirements.in
platformdirs==4.3.6
    # via textual
plotext==5.3.2
    # via -r requirements.in
pyaudio==0.2.14
    # via -r requirements.in
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
rich==13.9.4
    # via
    #   -r requirements.in
    #   textual
sniffio==1.3.1
    # via
    #   anyio
    #   openai
textual==1.0.0
    # via -r requirements.in
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   openai
    #   pydantic
    #   pydantic-core
    #   textual
uc-micro-py==1.0.3
    # via linkify-it-py
